Use size_t for array subscripts where possible

+++ src/iconv_ucs2.c
@@ -24,7 +24,7 @@
 struct iconv_ucs2
 {
     iconv_t cd;
-    unsigned pos;
+    size_t pos;
     unsigned add_bom;
 };
